prob_wheel {abtest}	R Documentation

Plot Probability Wheel

Description

Function for visualizing prior and posterior probabilities of the hypotheses as a probability wheel.

Usage


prob_wheel(x, type = "posterior")

Arguments

x
object of class "ab".

type
character indicating whether to plot a probability wheel visualizing the prior probabilities of the hypotheses (i.e., type = "prior") or the posterior probabilities of the hypotheses (i.e., type = "posterior"). The default is "posterior".

Author(s)

Quentin F. Gronau

Examples


# synthetic data
data <- list(y1 = 10, n1 = 28, y2 = 14, n2 = 26)

# Bayesian A/B test with default settings
ab <- ab_test(data = data)
print(ab)

# visualize prior probabilities of the hypotheses
prob_wheel(ab, type = "prior")

# visualize posterior probabilities of the hypotheses
prob_wheel(ab, type = "posterior")

[Package abtest version 0.2.0 Index]