seqdata {abtest}	R Documentation

Synthetic Sequential Data

Description

This data set contains synthetic sequential A/B data (500 observations in each of the two groups, where the observations are alternating between groups). y1 denotes the number of successes for the first group, n1 denotes the corresponding total number of observations for the first group. Similarly, y2 denotes the number of successes for the second group and n2 denotes the corresponding total number of observations for the second group.

Usage


seqdata

Format

A list with 4 elements.

Examples



data(seqdata)

# conduct Bayesian A/B test with default settings
ab <- ab_test(data = seqdata)
print(ab)

# produce sequential plot of posterior probabilities of the hypotheses
plot_sequential(ab, thin = 4)

# example of good width and height values for saving to file
cairo_pdf(file.path(tempdir(), "test_plot.pdf"),
          width = 530 / 72, height = 400 / 72)
plot_sequential(ab)
dev.off()


[Package abtest version 0.2.0 Index]